Fix another bunch of lintian warnings. Don't abuse (const char *)e_line, use a (char *)line instead. (maybe some leaks introduced? :)) Maybe a double free or too got fixed.

+++ common.h
@@ -36,7 +36,7 @@ const char *el_prompt_null(void);
 #endif
 const char *prompt_str(void);
 int malloc_check(void *);
-char *get_random_str(int, char);
+char *get_random_str(size_t, char);
 void quit(int);
